Take in the start and end dates for the requested data
For the response, we are going to filter it so that we only get back data that is within the timeframe that is useful for us. Also, there is no need to return back data about any other widget type other than 'meter'

TODO: Have to implement the same type of logic in the Checklist widget request.